+++ src/cljs/wombats_web_client/components/arena_table.cljs
@@ -1,5 +1,8 @@
 (ns wombats-web-client.components.arena-table
-  (:require [wombats-web-client.components.table :refer [table]]))
+  (:require [re-frame.core :as re-frame]
+            [wombats-web-client.components.table :refer [table]]
+            [wombats-web-client.components.modals.create-game-modal
+             :refer [create-game-modal]]))
 
 (defonce headers ["Name"
                   "Grid"
@@ -7,19 +10,40 @@
                   "Poison"
                   "Steel Barrier"
                   "Wood Barrier"
-                  "Zakano"])
+                  "Zakano"
+                  ""])
+
+
+(defn open-create-game-modal [arena-id]
+  (re-frame/dispatch [:set-modal {:fn #(create-game-modal arena-id)
+                                  :show-overlay true}]))
+
+(defn create-game-button [arena-id]
+  [:input.create-game {:type "button"
+                       :value "CREATE GAME"
+                       :on-click #(open-create-game-modal arena-id)}])
 
 (defn get-items-fn [row-data]
-  (let [name (:arena/name row-data)
-        width (:arena/width row-data)
-        height (:arena/height row-data)
-        food (:arena/food row-data)
-        poison (:arena/poison row-data)
-        steel (:arena/steel-walls row-data)
-        wood (:arena/wood-walls row-data)
-        zakano (:arena/zakano row-data)
-        dimension (str width "x" height)]
-    [name dimension food poison steel wood zakano]))
+  (let [{:keys [:arena/width
+                :arena/height
+                :arena/food
+                :arena/poison
+                :arena/steel-walls
+                :arena/wood-walls
+                :arena/zakano
+                :arena/id
+                :arena/name]} row-data
+                dimension (str width "x" height)
+                create-game (create-game-button id)]
+    [name
+     dimension
+     food
+     poison
+     steel-walls
+     wood-walls
+     zakano
+     create-game]))
 
 (defn arena-table [arenas]
-  [table "arena-table" headers @arenas get-items-fn])
+  [:div.arena-display
+   [table "arena-table" headers @arenas get-items-fn]])